+++ choreo_visual_tools/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package choreo_visual_tools
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-0.0.1 (2018-04-07)
+0.0.1 (2018-04-08)
 ------------------
-Support for input grasps (including end effector) visualization.
+Support for input geometry visualization.
+
+0.0.2 (2018-04-08)
+------------------
+Support for grasp + end effector visualization